Edit README.md
This reflects a change in direction philosophically. I want to separate
out the download functionality from the process of getting the file into
Anki at this point. It makes more sense to break this up into separate
functional units since for my own purposes it will be integrated into a
Keyboard Maestro workflow.