Remove Anki interaction features
This reflects a design decision to isolate downloading from processing
of the pronunciation files. Since for my own purposes, this will likely
only be used within the context of a Keyboard Maestro workflow, it makes
more sense to just use the simpler workflow features to finish the
processing there as required.